Changed the
try:
    for op in self.variators:
        op()
except TypeError

pattern.

It is more accurate to test whether a list or tuple is supplies as an argument.
In the case of checking for TypeErrors, *actual* TypeErrors in for instance a variator, than the error is hidden to the user, which is a lot harder to debug.
The pattern:

ifinstance(self.variators, (list,tuple):
    for op  in self.variators:
        op()
else:
  self.variators()
  
Switching according to data type is much cleaner and more precise in this context than catching errors.