Refactored the try/except blocks to use setdefault and get methods instead (except for a few unusual cases). It makes the code significantly shorter, replacing 5 lines with 1 in each occurrence.
Refactored the terminator to be an attribute of the EC, rather than being passed in to the evolve method. Now, only problem-specific things are passed to the evolve method (plus the ubiquitous args parameter).

Updated all the documentation to reflect the changes to the terminator (and to correct other oversights from changes made previously).

Consolidated the "TERMINATED DUE TO blah" by taking out all the print statements in each terminator and just printing in the _should_terminate using the name parameter. It makes the code cleaner and doesn't force us to continue to add print statements throughout the code.

Moved the version number from 0.5 to 0.6.